[Fusion] Renamed SyntaxException to FieldSelectionMapSyntaxException
